renamed Client to Browser
