+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -34,7 +34,7 @@
         "doctrine/cache": "~1.0",
         "fig/link-util": "^1.0",
         "symfony/asset": "~3.4|~4.0",
-        "symfony/browser-kit": "~3.4|~4.0",
+        "symfony/browser-kit": "^4.3",
         "symfony/console": "~3.4|~4.0",
         "symfony/css-selector": "~3.4|~4.0",
         "symfony/dom-crawler": "~3.4|~4.0",
@@ -67,6 +67,7 @@
         "phpdocumentor/type-resolver": "<0.2.1",
         "phpunit/phpunit": "<4.8.35|<5.4.3,>=5.0",
         "symfony/asset": "<3.4",
+        "symfony/browser-kit": "<4.3",
         "symfony/console": "<3.4",
         "symfony/dotenv": "<4.2",
         "symfony/form": "<4.3",
