+++ src/Symfony/Component/HttpKernel/composer.json
@@ -25,7 +25,7 @@
         "psr/log": "~1.0"
     },
     "require-dev": {
-        "symfony/browser-kit": "~3.4|~4.0",
+        "symfony/browser-kit": "^4.3",
         "symfony/config": "~3.4|~4.0",
         "symfony/console": "~3.4|~4.0",
         "symfony/css-selector": "~3.4|~4.0",
@@ -45,6 +45,7 @@
         "psr/log-implementation": "1.0"
     },
     "conflict": {
+        "symfony/browser-kit": "<4.3",
         "symfony/config": "<3.4",
         "symfony/dependency-injection": "<4.2",
         "symfony/translation": "<4.2",
